Run functions runtime test with thread mode
*Motivation*

Pulsar Functions supports both process and thread modes to run functions.
Currently the test case only test process mode.

*Changes*

Improve the existing test to support testing pulsar functions in thread mode
