+++ tests/integration/semantics/src/test/java/org/apache/pulsar/tests/integration/functions/runtime/PulsarFunctionsThreadRuntimeTest.java
@@ -16,39 +16,27 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.tests.integration.functions;
+package org.apache.pulsar.tests.integration.functions.runtime;
 
 import lombok.extern.slf4j.Slf4j;
-import org.apache.pulsar.tests.integration.functions.utils.CommandGenerator.Runtime;
 import org.apache.pulsar.tests.topologies.PulsarClusterTestBase;
 import org.testcontainers.containers.Container.ExecResult;
 import org.testng.annotations.BeforeClass;
-import org.testng.annotations.DataProvider;
 
 /**
- * A cluster to run pulsar functions for testing functions related features.
+ * Run the runtime test cases in thread mode.
  */
 @Slf4j
-public abstract class PulsarFunctionsTestBase extends PulsarClusterTestBase  {
-
-    public static final String EXCLAMATION_FUNC_CLASS =
-        "org.apache.pulsar.functions.api.examples.ExclamationFunction";
+public class PulsarFunctionsThreadRuntimeTest extends PulsarFunctionsRuntimeTest {
 
     @BeforeClass
     public static void setupCluster() throws Exception {
         PulsarClusterTestBase.setupCluster();
 
-        pulsarCluster.startFunctionWorkersWithProcessContainerFactory(1);
+        pulsarCluster.startFunctionWorkersWithThreadContainerFactory(1);
 
         ExecResult result = pulsarCluster.getAnyWorker().execCmd("cat", "/pulsar/conf/functions_worker.yml");
         log.info("Functions Worker Config : \n{}", result.getStdout());
-    }
-
-    @DataProvider(name = "FunctionRuntimes")
-    public static Object[][] functionRuntimes() {
-        return new Object[][] {
-            new Object[] { Runtime.JAVA }
-        };
     }
 
 }
